pyproject.toml
[project]
name = "alumni"
version = "0.1.0"
description = "WBGYM alumni"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["flask", "gunicorn", "flask-sqlalchemy", "flask-migrate", "dotenv", "Path", "datetime"]